+++ application/src/main/java/org/opentripplanner/gtfs/mapping/FareTransferRuleMapper.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Objects;
 import javax.annotation.Nullable;
 import org.opentripplanner.ext.fares.model.FareTransferRule;
+import org.opentripplanner.ext.fares.model.TimeLimitType;
 import org.opentripplanner.model.fare.FareProduct;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 
@@ -31,18 +32,18 @@ class FareTransferRuleMapper {
     var fareProductId = idFactory.createNullableId(rhs.getFareProductId());
     final var products = findFareProducts(fareProductId, rhs.getId());
 
-    Duration duration = null;
-    if (rhs.getDurationLimit() != MISSING_VALUE) {
-      duration = Duration.ofSeconds(rhs.getDurationLimit());
-    }
-    return FareTransferRule.of()
+    var builder = FareTransferRule.of()
       .withId(idFactory.createId(rhs.getId(), "fare transfer rule"))
       .withFromLegGroup(idFactory.createNullableId(rhs.getFromLegGroupId()))
       .withToLegGroup(idFactory.createNullableId(rhs.getToLegGroupId()))
       .withTransferCount(rhs.getTransferCount())
-      .withTimeLimit(duration)
-      .withFareProducts(products)
-      .build();
+      .withFareProducts(products);
+    if (rhs.getDurationLimit() != MISSING_VALUE) {
+      var duration = Duration.ofSeconds(rhs.getDurationLimit());
+      var type = mapLimitType(rhs.getDurationLimitType());
+      builder.withTimeLimit(type, duration);
+    }
+    return builder.build();
   }
 
   private Collection<FareProduct> findFareProducts(
@@ -60,5 +61,16 @@ class FareTransferRuleMapper {
       );
     }
     return products;
+  }
+
+  static TimeLimitType mapLimitType(int durationLimitType) {
+    return switch (durationLimitType) {
+      case 0 -> TimeLimitType.DEPARTURE_TO_ARRIVAL;
+      case 1 -> TimeLimitType.DEPARTURE_TO_DEPARTURE;
+      case 2, 3 -> throw new IllegalArgumentException(
+        "Duration limit type %s not implemented.".formatted(durationLimitType)
+      );
+      default -> throw new IllegalArgumentException("Valid duration limit type must be provided.");
+    };
   }
 }