+++ application/src/main/java/org/opentripplanner/standalone/api/OtpServerRequestContext.java
@@ -27,6 +27,7 @@ import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graphfinder.GraphFinder;
+import org.opentripplanner.routing.linking.LinkingContextFactory;
 import org.opentripplanner.routing.linking.VertexLinker;
 import org.opentripplanner.routing.via.ViaCoordinateTransferFactory;
 import org.opentripplanner.service.realtimevehicles.RealtimeVehicleService;
@@ -124,10 +125,10 @@ public interface OtpServerRequestContext {
 
   default GraphFinder graphFinder() {
     return GraphFinder.getInstance(
-      graph(),
-      vertexLinker(),
+      graph().hasStreets,
       transitService()::getRegularStop,
-      transitService()::findRegularStopsByBoundingBox
+      transitService()::findRegularStopsByBoundingBox,
+      linkingContextFactory()
     );
   }
 
@@ -183,4 +184,6 @@ public interface OtpServerRequestContext {
   FareService fareService();
 
   VertexLinker vertexLinker();
+
+  LinkingContextFactory linkingContextFactory();
 }