+++ application/src/main/java/org/opentripplanner/visualizer/ShowGraph.java
@@ -50,7 +50,6 @@ import org.opentripplanner.street.model.vertex.TransitPathwayNodeVertex;
 import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.model.vertex.VehicleParkingEntranceVertex;
 import org.opentripplanner.street.model.vertex.Vertex;
-import org.opentripplanner.street.search.TraverseMode;
 import org.opentripplanner.street.search.state.State;
 import processing.core.PApplet;
 import processing.core.PFont;
@@ -603,8 +602,6 @@ public class ShowGraph extends PApplet implements MouseWheelListener {
   private void drawGraphPath(GraphPath<State, Edge, Vertex> gp) {
     // draw edges in different colors according to mode
     for (State s : gp.states) {
-      TraverseMode mode = s.getBackMode();
-
       Edge e = s.getBackEdge();
       if (e == null) continue;