+++ raptor/src/test/java/org/opentripplanner/raptor/moduletests/F11_AccessWithRidesMultipleOptimalPathsTest.java
@@ -88,14 +88,11 @@ public class F11_AccessWithRidesMultipleOptimalPathsTest implements RaptorTestCo
    */
   static List<RaptorModuleTestCase> testCase3mWalkAccess() {
     var startFlexAccess = "Flex 11m 1x ~ C ";
-    var startWalkAndL1 = "A ~ BUS L1 0:02 0:10 ~ B ~ Walk 2m ~ C ";
     var endWalkAndL3 = "~ Walk 2m ~ D ~ BUS L3 0:16 0:22 ~ F ";
     var endL2AndWalk = "~ BUS L2 0:15 0:20 ~ E ~ Walk 3m ";
 
     // Min-Duration 19m - this is
     var flexTransferTransit = startFlexAccess + endWalkAndL3 + "[0:02 0:22 20m Tₓ1 C₁2_580]";
-    // Best duration 18m,
-    var transitAndTransit = startWalkAndL1 + endL2AndWalk + "[0:02 0:23 21m Tₓ1]";
     // Min-Duration 19m
     var flexAndTransit = startFlexAccess + endL2AndWalk + "[0:03 0:23 20m Tₓ1 C₁2_640]";
 
@@ -121,11 +118,7 @@ public class F11_AccessWithRidesMultipleOptimalPathsTest implements RaptorTestCo
   static List<RaptorModuleTestCase> testCase1mWalkAccess() {
     var startFlexAccess = "Flex 11m 1x ~ C ";
     var startWalkAndL1 = "A ~ BUS L1 0:02 0:10 ~ B ~ Walk 2m ~ C ";
-    var endWalkAndL3 = "~ Walk 2m ~ D ~ BUS L3 0:16 0:22 ~ F ";
     var endL2AndWalk = "~ BUS L2 0:15 0:20 ~ E ~ Walk 1m ";
-
-    // Min-Duration 19m - this is
-    var flexTransferTransit = startFlexAccess + endWalkAndL3 + "[0:02 0:22 20m Tₓ1 C₁2_580]";
     // Best duration 16m,
     var transitAndTransit = startWalkAndL1 + endL2AndWalk + "[0:02 0:21 19m Tₓ1]";
     // Min-Duration 17m