+++ doc/user/Changelog.md
@@ -34,6 +34,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Fix duplicates in paging due to cost truncation in token [#6973](https://github.com/opentripplanner/OpenTripPlanner/pull/6973)
 - Added transit service validity dates and CPU cores to GraphQL ServerInfo and transitInfo [#6996](https://github.com/opentripplanner/OpenTripPlanner/pull/6996)
 - SIRI-ET updater via MQTT [#6851](https://github.com/opentripplanner/OpenTripPlanner/pull/6851)
+- Support for ETag and If-Not-Modified in GBFS updater [#6963](https://github.com/opentripplanner/OpenTripPlanner/pull/6963)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.8.1 (2025-09-22)